# docs/OPENAPI.yaml
openapi: 3.1.0
info:
  title: Quantoom Root Platform API
  description: Unified API for Quantum Computing Simulation, AI Copilot RAG, Ethical Cyber Range, and Community Hub.
  version: 1.0.0
servers:
  - url: https://api.quantoomroot.io/api/v1
    description: Production Server
  - url: http://localhost:4000/api/v1
    description: Local Development Server
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, username, password]
              properties:
                email: { type: string, format: email }
                username: { type: string }
                password: { type: string }
                primaryDomain: { type: string, enum: [QUANTUM, AI, CYBERSECURITY, CONVERGENCE] }
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Validation error
  /auth/login:
    post:
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: { type: string }
                password: { type: string }
                mfaCode: { type: string }
      responses:
        '200':
          description: Authentication successful
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT